refactor(#792): split audio_alloc_invariant_tests.rs (634->231 base + real-rig file)
Base keeps the #[global_allocator] instrumentation + fixtures (pub(super)); #670/#781
real-block tests move to audio_alloc_real_rig_tests.rs (they share p670_* fixtures).
Off debt. Alloc invariant: 9 pass with --ignored --test-threads=1.
