+++ scripts/validate.sh
@@ -52,7 +52,6 @@ crates/adapter-gui/src/chain_row_wiring.rs
 crates/adapter-gui/src/compact_chain_block_handlers.rs
 crates/adapter-gui/src/desktop_app.rs
 crates/application/src/command.rs
-crates/engine/src/audio_alloc_invariant_tests.rs
 crates/engine/src/runtime_tests.rs
 crates/engine/tests/issue_670_beat_it_real_rig.rs
 crates/infra-cpal/src/controller_live_edit_replicates_user_report_tests.rs
